custom borders and added icons

+++ app/src/main/java/com/crossdevelop/cryptocoincompose/common/models/CoinDetailCurrentPrice.kt
@@ -2,7 +2,7 @@ package com.crossdevelop.cryptocoincompose.common.models
 
 
 data class CoinDetailCurrentPrice(
-    val usd: Float,
-    val eth: Float,
-    val btc: Float
+    val usd: Double,
+    val eth: Double,
+    val btc: Double
 )
